@@ -16,26 +16,26 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^9.1.13",
-    "@angular/common": "^9.1.13",
-    "@angular/compiler": "^9.1.13",
-    "@angular/core": "^9.1.13",
-    "@angular/forms": "^9.1.13",
-    "@angular/localize": "^11.2.10",
-    "@angular/platform-browser": "^9.1.13",
-    "@angular/platform-browser-dynamic": "^9.1.13",
-    "@angular/router": "^9.1.13",
-    "@fortawesome/angular-fontawesome": "0.7.0",
+    "@angular/animations": "^11.2.12",
+    "@angular/common": "^11.2.12",
+    "@angular/compiler": "^11.2.12",
+    "@angular/core": "^11.2.12",
+    "@angular/forms": "^11.2.12",
+    "@angular/localize": "^11.2.12",
+    "@angular/platform-browser": "^11.2.12",
+    "@angular/platform-browser-dynamic": "^11.2.12",
+    "@angular/router": "^11.2.12",
+    "@fortawesome/angular-fontawesome": "0.8.2",
     "@fortawesome/fontawesome-svg-core": "^1.2.35",
     "@fortawesome/free-regular-svg-icons": "^5.14.0",
     "@fortawesome/free-solid-svg-icons": "^5.14.0",
-    "@ng-bootstrap/ng-bootstrap": "^6.2.0",
-    "@schematics/angular": "^9.1.15",
-    "@sentry/browser": "^5.23.0",
+    "@ng-bootstrap/ng-bootstrap": "^9.1.0",
+    "@schematics/angular": "^11.2.11",
+    "@sentry/browser": "^6.3.5",
     "@types/file-saver": "^2.0.2",
-    "@types/jwt-decode": "^2.2.1",
+    "@types/jwt-decode": "^3.1.0",
     "add": "^2.0.6",
-    "core-js": "^3.6.5",
+    "core-js": "^3.11.1",
     "file-saver": "^2.0.5",
     "file-type": "^16.1.0",
     "fs": "^0.0.2",
@@ -45,32 +45,32 @@
     "ngx-countup": "7.3.3",
     "ngx-matomo": "^0.1.4",
     "ngx-print": "1.2.0-beta.6",
-    "ngx-toastr": "^12.1.0",
+    "ngx-toastr": "^13.2.1",
     "path": "^0.12.7",
     "raven-js": "^3.27.2",
-    "rxjs": "^6.6.7",
+    "rxjs": "^7.0.0",
     "smoothscroll-polyfill": "^0.4.4",
-    "tslib": "^2.0.3",
+    "tslib": "^2.0.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.901.15",
-    "@angular/cli": "^9.1.15",
-    "@angular/compiler-cli": "^9.1.13",
-    "@angular/language-service": "^9.1.13",
-    "@schematics/update": "^0.901.15",
+    "@angular-devkit/build-angular": "^0.1102.11",
+    "@angular/cli": "^11.2.11",
+    "@angular/compiler-cli": "^11.2.12",
+    "@angular/language-service": "^11.2.12",
+    "@schematics/update": "^0.1102.11",
     "@socialgouv/tslint-config-recommended": "^0.35.0",
     "@types/file-saver": "^2.0.2",
-    "@types/jest": "^26.0.22",
-    "@types/node": "^8.9.5",
-    "codelyzer": "^6.0.1",
+    "@types/jest": "^26.0.23",
+    "@types/node": "^15.0.1",
+    "codelyzer": "^6.0.0",
     "eslint-plugin-import": "^2.22.1",
     "jest": "^26.4.2",
     "jest-preset-angular": "^8.3.2",
     "prettier": "^2.1.2",
-    "ts-jest": "^26.3.0",
-    "ts-node": "~9.0.0",
-    "tslint": "^6.1.3",
-    "typescript": "3.8.3"
+    "ts-jest": "^26.5.5",
+    "ts-node": "~9.1.1",
+    "tslint": "~6.1.0",
+    "typescript": "4.1.5"
   }
 }
