@@ -176,6 +176,6 @@ export const routes: Routes = [
 ];
 @NgModule({
   exports: [RouterModule],
-  imports: [RouterModule.forRoot(routes)],
+  imports: [RouterModule.forRoot(routes, { relativeLinkResolution: 'legacy' })],
 })
 export class AppRoutingModule {}
