@@ -1,7 +1,8 @@
-import { ErrorHandler } from "@angular/core";
+import { ErrorHandler, Injectable } from "@angular/core";
 
 import * as Sentry from "@sentry/browser";
 
+@Injectable()
 export class SentryErrorHandler implements ErrorHandler {
   handleError(err: any): void {
     Sentry.captureException(new Error(err));
