@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from "@angular/core/testing";
+import { ComponentFixture, TestBed, waitForAsync } from "@angular/core/testing";
 
 import { MentionsLegalesComponent } from "./mentions-legales.component";
 
@@ -6,7 +6,7 @@ describe("MentionsLegalesComponent", () => {
   let component: MentionsLegalesComponent;
   let fixture: ComponentFixture<MentionsLegalesComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [MentionsLegalesComponent],
     }).compileComponents();
