@@ -15,23 +15,25 @@ describe("NavbarComponent", () => {
   let component: NavbarComponent;
   let fixture: ComponentFixture<NavbarComponent>;
 
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [NavbarComponent],
-      imports: [
-        CommonModule,
-        SharedModule,
-        BrowserAnimationsModule,
-        FontAwesomeModule,
-        RouterTestingModule,
-        HttpClientModule,
-        HttpClientTestingModule,
-      ],
-      providers: [{ provide: APP_BASE_HREF, useValue: "/" }],
+  beforeEach(
+    waitForAsync(() => {
+      TestBed.configureTestingModule({
+        declarations: [NavbarComponent],
+        imports: [
+          CommonModule,
+          SharedModule,
+          BrowserAnimationsModule,
+          FontAwesomeModule,
+          RouterTestingModule,
+          HttpClientModule,
+          HttpClientTestingModule,
+        ],
+        providers: [{ provide: APP_BASE_HREF, useValue: "/" }],
 
-      schemas: [CUSTOM_ELEMENTS_SCHEMA, NO_ERRORS_SCHEMA],
-    }).compileComponents();
-  }));
+        schemas: [CUSTOM_ELEMENTS_SCHEMA, NO_ERRORS_SCHEMA],
+      }).compileComponents();
+    })
+  );
 
   beforeEach(() => {
     fixture = TestBed.createComponent(NavbarComponent);
