@@ -1,6 +1,6 @@
 import { APP_BASE_HREF } from "@angular/common";
 import { CUSTOM_ELEMENTS_SCHEMA } from "@angular/core";
-import { async, ComponentFixture, TestBed } from "@angular/core/testing";
+import { ComponentFixture, TestBed, waitForAsync } from "@angular/core/testing";
 import { RouterModule } from "@angular/router";
 import { NgbModule } from "@ng-bootstrap/ng-bootstrap";
 import { LoadingComponent } from "./loading.component";
@@ -11,21 +11,21 @@ describe("LoadingComponent", () => {
   let fixture: ComponentFixture<LoadingComponent>;
   let service: LoadingService;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [LoadingComponent],
-      imports: [NgbModule, RouterModule.forRoot([])],
+      imports: [NgbModule, RouterModule.forRoot([], { relativeLinkResolution: 'legacy' })],
       providers: [{ provide: APP_BASE_HREF, useValue: "/" }],
       schemas: [CUSTOM_ELEMENTS_SCHEMA],
     }).compileComponents();
 
     fixture = TestBed.createComponent(LoadingComponent);
     component = fixture.componentInstance;
-    service = TestBed.get(LoadingService);
+    service = TestBed.inject(LoadingService);
     fixture.detectChanges();
   }));
 
-  it("should create", async(() => {
+  it("should create", waitForAsync(() => {
     expect(component).toBeTruthy();
     expect(component.loading).toBeFalsy();
 
