@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from "@angular/core/testing";
+import { ComponentFixture, TestBed, waitForAsync } from "@angular/core/testing";
 
 import { APP_BASE_HREF } from "@angular/common";
 
@@ -14,7 +14,7 @@ describe("StructuresFormComponent", () => {
   let component: StructuresFormComponent;
   let fixture: ComponentFixture<StructuresFormComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [StructuresFormComponent],
       imports: [
