@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from "@angular/core/testing";
+import { ComponentFixture, TestBed, waitForAsync } from "@angular/core/testing";
 
 import { APP_BASE_HREF } from "@angular/common";
 import { HttpClientModule } from "@angular/common/http";
@@ -19,7 +19,7 @@ describe("StructuresSearchComponent", () => {
   let component: StructuresSearchComponent;
   let fixture: ComponentFixture<StructuresSearchComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [StructuresSearchComponent],
       imports: [
