@@ -44,8 +44,8 @@ import { StructureService } from "./services/structure.service";
     SharedModule,
     BrowserAnimationsModule,
     FontAwesomeModule,
-    RouterModule.forRoot([]),
-    ToastrModule.forRoot(),
+    RouterModule.forRoot([], { relativeLinkResolution: "legacy" }),
+    ToastrModule.forRoot({}),
     HttpClientModule,
     NgbModule,
     FormsModule,
