@@ -1,6 +1,6 @@
 import { APP_BASE_HREF } from "@angular/common";
 import { CUSTOM_ELEMENTS_SCHEMA } from "@angular/core";
-import { async, ComponentFixture, TestBed } from "@angular/core/testing";
+import { ComponentFixture, TestBed, waitForAsync } from "@angular/core/testing";
 import { MatomoInjector, MatomoTracker } from "ngx-matomo";
 import { UsagersModule } from "src/app/modules/usagers/usagers.module";
 import { usagerValideMock } from "../../../../../../../_common/mocks/usager.mock";
@@ -11,7 +11,7 @@ describe("RdvComponent", () => {
   let component: RdvComponent;
   let fixture: ComponentFixture<RdvComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [],
       imports: [UsagersModule],
