@@ -3,7 +3,7 @@ import { HttpClientModule } from "@angular/common/http";
 import { HttpClientTestingModule } from "@angular/common/http/testing";
 import { global } from "@angular/compiler/src/util";
 import { CUSTOM_ELEMENTS_SCHEMA } from "@angular/core";
-import { async, ComponentFixture, TestBed } from "@angular/core/testing";
+import { ComponentFixture, TestBed, waitForAsync } from "@angular/core/testing";
 import { FormsModule, ReactiveFormsModule } from "@angular/forms";
 import { RouterTestingModule } from "@angular/router/testing";
 import { NgbModule } from "@ng-bootstrap/ng-bootstrap";
@@ -61,7 +61,7 @@ describe("ManageUsagersComponent", () => {
     expect(component).toBeTruthy();
   });
 
-  it("3. Reset Filters", async(() => {
+  it("3. Reset Filters", waitForAsync(() => {
     component.resetFilters();
 
     expect(component.filters).toEqual({
