@@ -2,7 +2,7 @@ import { APP_BASE_HREF } from "@angular/common";
 import { HttpClientModule } from "@angular/common/http";
 import { HttpClientTestingModule } from "@angular/common/http/testing";
 import { CUSTOM_ELEMENTS_SCHEMA, NO_ERRORS_SCHEMA } from "@angular/core";
-import { async, ComponentFixture, TestBed } from "@angular/core/testing";
+import { ComponentFixture, TestBed, waitForAsync } from "@angular/core/testing";
 import { FormsModule, ReactiveFormsModule } from "@angular/forms";
 import { RouterTestingModule } from "@angular/router/testing";
 import { ToastrModule } from "ngx-toastr";
@@ -12,7 +12,7 @@ describe("UserProfilComponent", () => {
   let component: UserProfilComponent;
   let fixture: ComponentFixture<UserProfilComponent>;
 
-  beforeAll(async(() => {
+  beforeAll(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [UserProfilComponent],
       imports: [
