@@ -40,8 +40,8 @@ import { SharedModule } from "../shared/shared.module";
     SharedModule,
     BrowserAnimationsModule,
     FontAwesomeModule,
-    RouterModule.forRoot([]),
-    ToastrModule.forRoot(),
+    RouterModule.forRoot([], { relativeLinkResolution: "legacy" }),
+    ToastrModule.forRoot({}),
     HttpClientModule,
     NgbModule,
     FormsModule,
